+++ src/QueryCat.Backend/Formatters/TextTableOutput.cs
@@ -2,6 +2,7 @@ using Microsoft.Extensions.Logging;
 using QueryCat.Backend.Relational;
 using QueryCat.Backend.Relational.Iterators;
 using QueryCat.Backend.Storage;
+using QueryCat.Backend.Types;
 
 namespace QueryCat.Backend.Formatters;
 
@@ -30,6 +31,7 @@ public sealed class TextTableOutput : RowsOutput, IDisposable
     private readonly bool _hasHeader;
     private readonly string _separator;
     private readonly string _separatorWithSpace;
+    private readonly string _floatNumberFormat;
 
     private readonly Action _onInit;
     private readonly Action<Row> _onWrite;
@@ -43,11 +45,16 @@ public sealed class TextTableOutput : RowsOutput, IDisposable
     /// </summary>
     public string Separator => _separator;
 
-    public TextTableOutput(Stream stream, bool hasHeader = true, string? separator = null,
-        Style style = Style.Table)
+    public TextTableOutput(
+        Stream stream,
+        bool hasHeader = true,
+        string? separator = null,
+        Style style = Style.Table,
+        string? floatNumberFormat = null)
     {
         _stream = stream;
         _hasHeader = hasHeader;
+        _floatNumberFormat = floatNumberFormat ?? VariantValue.FloatNumberFormat;
 
         if (style == Style.Card)
         {
@@ -108,13 +115,6 @@ public sealed class TextTableOutput : RowsOutput, IDisposable
 
         _onInit.Invoke();
         _streamWriter.Flush();
-    }
-
-    /// <inheritdoc />
-    public void Dispose()
-    {
-        Close();
-        _stream.Dispose();
     }
 
     #region Table
@@ -198,7 +198,7 @@ public sealed class TextTableOutput : RowsOutput, IDisposable
                 writeCount += _separatorWithSpace.Length;
             }
             var value = row[i];
-            var valueString = value.ToString();
+            var valueString = ToStringWithFormat(value);
             var padding = _columnsLengths[i];
             var exceed = _totalMaxLineLength[i] - writeCount - _columnsLengths[i] - 1;
             if (exceed < 0)
@@ -254,8 +254,7 @@ public sealed class TextTableOutput : RowsOutput, IDisposable
             {
                 continue;
             }
-            var value = row[i];
-            var valueString = value.ToString();
+            var valueString = ToStringWithFormat(row[i]);
             _streamWriter.Write(_separator);
             _streamWriter.Write(valueString);
         }
@@ -286,12 +285,29 @@ public sealed class TextTableOutput : RowsOutput, IDisposable
             {
                 continue;
             }
-            var value = row[i];
+            var valueString = ToStringWithFormat(row[i]);
             _streamWriter.Write(row.Columns[i].FullName.PadLeft(_maxColumnNameWidth));
-            _streamWriter.Write($" {_separator} {value}\n");
+            _streamWriter.Write($" {_separator} {valueString}\n");
         }
         _streamWriter.WriteLine();
     }
 
     #endregion
+
+    private string ToStringWithFormat(in VariantValue value)
+    {
+        var type = value.GetInternalType();
+        if (type == DataType.Float || type == DataType.Numeric)
+        {
+            return value.ToString(_floatNumberFormat);
+        }
+        return value.ToString();
+    }
+
+    /// <inheritdoc />
+    public void Dispose()
+    {
+        Close();
+        _stream.Dispose();
+    }
 }