+++ src/QueryCat.Cli/Commands/Options/QueryOptionsBinder.cs
@@ -1,6 +1,7 @@
 using System.CommandLine;
 using System.CommandLine.Binding;
 using QueryCat.Backend.Formatters;
+using QueryCat.Backend.Types;
 
 namespace QueryCat.Cli.Commands.Options;
 
@@ -16,6 +17,7 @@ internal class QueryOptionsBinder : BinderBase<QueryOptions>
     private readonly Option<string?> _columnsSeparatorOption;
     private readonly Option<bool> _disableCacheOption;
     private readonly Option<bool> _noHeaderOption;
+    private readonly Option<string> _floatNumberOption;
 
     public QueryOptionsBinder(
         Option<int> maxErrorsOption,
@@ -26,7 +28,8 @@ internal class QueryOptionsBinder : BinderBase<QueryOptions>
         Option<int> analyzeRowsOption,
         Option<string?> columnsSeparatorOption,
         Option<bool> disableCacheOption,
-        Option<bool> noHeaderOption)
+        Option<bool> noHeaderOption,
+        Option<string> floatNumberOption)
     {
         _maxErrorsOption = maxErrorsOption;
         _statisticOption = statisticOption;
@@ -45,6 +48,7 @@ internal class QueryOptionsBinder : BinderBase<QueryOptions>
         _columnsSeparatorOption = columnsSeparatorOption;
         _disableCacheOption = disableCacheOption;
         _noHeaderOption = noHeaderOption;
+        _floatNumberOption = floatNumberOption;
     }
 
     /// <inheritdoc />
@@ -62,6 +66,8 @@ internal class QueryOptionsBinder : BinderBase<QueryOptions>
             ColumnsSeparator = bindingContext.ParseResult.GetValueForOption(_columnsSeparatorOption),
             DisableCache = bindingContext.ParseResult.GetValueForOption(_disableCacheOption),
             NoHeader = bindingContext.ParseResult.GetValueForOption(_noHeaderOption),
+            FloatNumberFormat = bindingContext.ParseResult.GetValueForOption(_floatNumberOption)
+                ?? VariantValue.FloatNumberFormat,
         };
     }
 }