+++ src/QueryCat.Cli/Commands/QueryCommand.cs
@@ -42,6 +42,9 @@ internal class QueryCommand : BaseQueryCommand
             description: "Do not use memory cache for sub-queries.");
         var noHeaderOption = new Option<bool>("--no-header",
             description: "Do not render header.");
+        var floatNumberOption = new Option<string>("--float-format",
+            description: "Float numbers format.",
+            getDefaultValue: () => "F");
 
         this.AddOption(maxErrorsOption);
         this.AddOption(statisticOption);
@@ -53,6 +56,7 @@ internal class QueryCommand : BaseQueryCommand
         this.AddOption(columnsSeparatorOption);
         this.AddOption(disableCacheOption);
         this.AddOption(noHeaderOption);
+        this.AddOption(floatNumberOption);
         this.SetHandler((applicationOptions, query, variables, files, queryOptions) =>
         {
             applicationOptions.InitializeLogger();
@@ -60,7 +64,8 @@ internal class QueryCommand : BaseQueryCommand
                 stream: StandardInputOutput.GetConsoleOutput(),
                 hasHeader: !queryOptions.NoHeader,
                 separator: queryOptions.ColumnsSeparator,
-                style: queryOptions.OutputStyle);
+                style: queryOptions.OutputStyle,
+                floatNumberFormat: queryOptions.FloatNumberFormat);
             var options = new ExecutionOptions
             {
                 AddRowNumberColumn = queryOptions.RowNumberOption,
@@ -103,7 +108,8 @@ internal class QueryCommand : BaseQueryCommand
                 analyzeRowsOption,
                 columnsSeparatorOption,
                 disableCacheOption,
-                noHeaderOption)
+                noHeaderOption,
+                floatNumberOption)
             );
     }
 }